Windows port: conv: fix narrowing conversions in braced initializers
- conv_transpose2d.cpp: static_cast int/uint32_t mismatches feeding
  std::max<int>({...}) and std::array<uint32_t,4> initialization
- conv2d_op_sharded_program_factory.cpp: static_cast CoreCoord/size_t
  fields to the expected int32_t/uint32_t types in braced initializers
- conv2d_op_width_sharded_program_factory.cpp: static_cast size_t
  core-grid dimension into a uint32_t runtime-args initializer list
- prepare_conv2d_weights.cpp: static_cast int loop indices to uint32_t
  in SmallVector braced initializers passed to compute_flat_indices
- sliding_window.hpp (shared header, needed by conv): static_cast the
  int 0 default member initializers for TensorMemoryLayout and
  ShardOrientation, which MSVC rejects as an enum-class narrowing
  conversion (GCC/Clang normally accept enum{0} without a cast)

+++ ttnn/cpp/ttnn/operations/conv/conv2d/device/conv2d_op_width_sharded_program_factory.cpp
@@ -636,7 +636,7 @@ tt::tt_metal::ProgramDescriptor build_program_descriptor(
         std::vector<uint32_t> rt_args = {
             core_x,
             core_y,
-            full_core_grid.x,  // num_cores_x
+            static_cast<uint32_t>(full_core_grid.x),  // num_cores_x
         };
 
         // Mcast X Lookup table
